implement more OOP
