+++ src/main/java/command/Command.java
@@ -1,0 +1,5 @@
+package command;
+
+public abstract class Command {
+    public abstract void execute(String command);
+}
