+++ src/main/java/command/DeleteCommand.java
@@ -1,0 +1,4 @@
+package command;
+
+public class DeleteCommand {
+}
