+++ src/main/java/command/EventCommand.java
@@ -1,0 +1,4 @@
+package command;
+
+public class EventCommand {
+}
