+++ src/main/java/command/ListCommand.java
@@ -1,0 +1,4 @@
+package command;
+
+public class ListCommand {
+}
