+++ src/main/java/command/MarkCommand.java
@@ -1,0 +1,4 @@
+package command;
+
+public class MarkCommand {
+}
