+++ src/main/java/exception/UnknownCommandException.java
@@ -1,5 +1,4 @@
 package exception;
-
 public class UnknownCommandException extends DukeException {
     public UnknownCommandException(String e) {
         super(e);
