+++ src/main/java/parser/Parser.java
@@ -1,0 +1,5 @@
+package parser;
+
+public class Parser {
+
+}
