+++ src/main/java/storage/Storage.java
@@ -1,0 +1,7 @@
+package storage;
+
+import java.io.File;
+
+public class Storage {
+    private File file;
+}
